feat(store): Capture sizes of stored event and attachment
